requirements.txt
# Core Dependencies
PyQt6==6.7.0
PyQt6-Qt6==6.7.0
PyQt6-sip==13.8.0

# Screen Capture & Detection
mss==9.0.1
screeninfo==0.8.1
Pillow==10.1.0

# OCR - PaddleOCR
paddleocr==2.7.0.3
paddlepaddle==3.0.0

# Computer Vision
opencv-python==4.6.0.66
opencv-contrib-python==4.6.0.66
numpy==1.26.4

# Translation APIs (NOVAS - GRATUITAS)
deep-translator>=1.11.4
requests>=2.31.0

# Database & Caching
SQLAlchemy==2.0.23

# Configuration
PyYAML==6.0.1
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# PaddleOCR dependencies
beautifulsoup4
cython
fire>=0.3.0
lxml
openpyxl
pdf2docx
premailer
python-docx
PyMuPDF==1.20.0

# Testing
pytest==7.4.3
